+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 `@/web-server`
 - Adds support for automatic conversion of boolean values when using `@QueryParam()` on a parameter of type `boolean`.
 The values `''`, `'no'`, `'0'`, `'false'`, and `'off'` produce `false`, all other values produce `true`.
+- Adds support for receiving the ID for the request from a request header (ie `X-Trace` or so). Not enabled by default. Use `requestIdHeader` option when configuring the web server to enable this functionality.
+- Fixes an issue where request ID was wastefully generated twice
 
 # 🚀 3.0.0-rc.4